Python 3 fix for the user function docstring creation in the prompt UI mode.
Again this is the problem of a division now returning a float rather than an int.

+++ prompt/uf_docstring.py
@@ -142,7 +142,7 @@ def create_table(label):
         # Loop.
         while True:
             # The average column width.
-            ave_width = free_space_wrap / num_cols_wrap
+            ave_width = int(free_space_wrap / num_cols_wrap)
 
             # Debugging printout.
             if status.debug:
